requirements.txt
python-dotenv
python-telegram-bot
torch
torchvision
torchaudio
requests
pyyaml
scipy
pandas
matplotlib
sympy
cryptography
Pillow
IPython
black
